backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6

